--- pkg/cache-proxy/README.md
@@ -1,3 +1,0 @@
-# Cache Proxy
-
-Runs a local server that is able to take requests from Buildkit's Remote Cache GHA backend and proxy them to work with WarpCache.
